pyproject.toml
[tool.poetry]
name = "cb-stream"
version = "0.1.0"
description = ""
authors = ["Marko Vukovic <mvkvc.git@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
python-dotenv = "^1.0.0"
websockets = "^11.0"
azure-eventhub = "^5.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
